Equation3DShapes (Abstract)
Its a class meant to organize some of the methods I reused in making classes similar to Desert. Not particularly useful, but necessary for Desert (unless you remove it as a parent-class in the Desert code itself). I intend to use it to generalize some of the methods used in Desert to make new classes like Desert easier to write.