Desert Editor
Running this applet allows you to play around with the parameters that specify the shape of desert. Default size of a Desert is 10 squares, but running a smaller size is recommended if you don't want to get a heap space error while making finer details. The UI is basic (understatement), but that's because I wrote Editor just so I could play around with the numbers and decide what configuration looks most "desert"-like.